refactor: book의 controller 패키지를 presentation으로 리네임
